+++ CHANGELOG.md
@@ -2,10 +2,15 @@
 
 All notable changes to `Laravel Wallet` will be documented in this file
 
+## 1.0.2 - 2023-09-20
+
+### Changed
+- Update `display` method on Money DTO.
+
 ## 1.0.1 - 2023-09-19
 
 ### Fixed
-- Declaration of Money cast not compatible
+- Declaration of Money cast not compatible.
 
 ## 1.0.0 - 2023-09-19